Rename reconfigure entry point & remove separate reconfigure methods
