--- content/automate/ManageIQ/Service/Generic/StateMachines/GenericLifecycle.class/__methods__/reconfigure_preprocess.yaml
@@ -1,14 +1,0 @@
----
-object_type: method
-version: 1.0
-object:
-  attributes:
-    name: reconfigure_preprocess
-    display_name:
-    description:
-    scope: instance
-    language: ruby
-    location: inline
-    embedded_methods:
-    - "/Service/Generic/StateMachines/Utils/util_object"
-  inputs: []
