--- content/automate/ManageIQ/Service/Generic/StateMachines/GenericLifecycle.class/__methods__/reconfigure_start.yaml
@@ -1,15 +1,0 @@
----
-object_type: method
-version: 1.0
-object:
-  attributes:
-    name: reconfigure_start
-    display_name:
-    description:
-    scope: instance
-    language: ruby
-    location: inline
-    embedded_methods:
-    - "/System/CommonMethods/Utils/log_object"
-    - "/Service/Generic/StateMachines/Utils/util_object"
-  inputs: []
