+++ content/automate/ManageIQ/Service/Generic/StateMachines/GenericLifecycle.class/reconfigure.yaml
@@ -1,0 +1,47 @@
+---
+object_type: instance
+version: 1.0
+object:
+  attributes:
+    display_name:
+    name: reconfigure
+    inherits:
+    description:
+    relative_path: Service/Generic/StateMachines/GenericLifecycle/reconfigure
+  fields:
+  - start:
+      value: METHOD::start
+      on_entry: update_status(status => 'Processing Reconfigure Start')
+      on_exit: update_status(status => 'Processed Reconfigure Start')
+      on_error: update_status(status => 'Error Processing Reconfigure Start')
+  - pre4:
+      value: METHOD::preprocess
+      on_entry: update_status(status => 'Processing Reconfigure Preprocess')
+      on_exit: update_status(status => 'Processed Reconfigure Preprocess')
+      on_error: update_status(status => 'Error Processing Reconfigure Preprocess')
+  - execute:
+      value: METHOD::execute
+      on_entry: update_status(status => 'Starting execute')
+      on_exit: update_status(status => 'Ending execute')
+      on_error: update_status(status => 'Error Processing execute')
+  - check_completed:
+      value: METHOD::check_completed
+      on_exit: update_status(status => 'Processed check_completed')
+      on_error: update_status(status => 'Error Processing check_completed')
+  - refresh:
+      value: METHOD::refresh
+      on_entry: update_status(status => 'Processing refresh')
+      on_exit: update_status(status => 'Processed refresh')
+      on_error: update_status(status => 'Error Processing refresh')
+  - check_refreshed:
+      value: METHOD::check_refreshed
+      on_exit: update_status(status => 'Processed check_refreshed')
+      on_error: update_status(status => 'Error Processing check_refreshed')
+  - post1:
+      value: METHOD::postprocess
+      on_entry: update_status(status => 'Processing Postprocess')
+      on_exit: update_status(status => 'Processed Postprocess')
+      on_error: update_status(status => 'Error Processing Postprocess')
+  - finish:
+      value: METHOD::finish
+      on_exit: update_status(status => "${/#service_action} Completed Successfully")
